Mcmc samples

Helper function to get the mcmc samples

get_mcmc_samples(
  data,
  id_application,
  id_voter,
  grade_variable,
  model = NULL,
  path_to_jags_model = NULL,
  n_chains = 2,
  n_iter = 10000,
  n_burnin = 5000,
  id_section = NULL,
  theta_name = "application_intercept",
  tau_name = "tau_application",
  sigma_name = "sigma",
  tau_voter_name = "tau_voter",
  tau_section_name = NULL,
  rank_theta_name = "rank_theta",
  other_variables = NULL,
  seed = 1991,
  quiet = FALSE,
  dont_bind = FALSE
)

Arguments

data
long data frame with all information as in the jags model defined below.

id_application
the name of the application variable in the data

id_voter
the name of the voter variable in the data (default = NULL)

grade_variable
the name of the outcome variable

model
the model definition as character. (default = NULL)

path_to_jags_model
the path to the jags txt file, if null, the default model is used. (default = NULL)

n_chains
number of chains for the JAGS sampler. (default = 4)

n_iter
how many iterations used in the JAGS sampler? (default = 5000)

n_burnin
number of burnin iterations

id_section
name of the section

theta_name
the name of the application identifier in the JAGS model. (default = application_intercept")

tau_name
name of the tau in the jags model, being the standard error of the application effects. This variable is needed for the computation of the rankability. (default = tau_application)

sigma_name
name of the standard deviation of the full model. (default = sigma)

tau_voter_name
name of the standard error of the voter effect. This variable is needed for the computation of the rankability (default = tau_voter)

tau_section_name
name of the standard error of the section effect, if needed

rank_theta_name
the name of the rank of theta in the JAGS model (default = rank_theta)

other_variables
are there other variables we would like to extract? (NULL by default)

seed
set a seed for the JAGS model (default = 1991)

quiet
if the default model is used this function generates a warning. if quiet = TRUE, this warning is not shown

dont_bind
if TRUE the different chains are not pooled, and the MCMC object is returned as it is.

Value

matrix with the samples for all parameters defined in the model.

Examples

data_panel1 <- get_mock_data() %>%
     filter(panel == "p1")

#> Error in get_mock_data() %>% filter(panel == "p1"): konnte Funktion "%>%" nicht finden

mcmc_samples <- get_mcmc_samples(data = data_panel1,
                                 id_application = "application",
                                 id_voter = "voter",
                                 grade_variable = "num_grade")

#> Error in c(id_application, id_voter, grade_variable, id_section) %in%     names(data): Objekt 'data_panel1' nicht gefunden